move mocks into package
